add mdx-pages plugin
