+++ packages/documentation/gatsby-config.js
@@ -58,5 +58,6 @@ module.exports = {
     `gatsby-transformer-react-docgen`,
     `gatsby-transformer-remark`,
     `github-api-pages`,
+    `mdx-pages`,
   ],
 }
