+++ packages/documentation/plugins/mdx-pages/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "mdx-pages",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
